{12/12/2018} Alterado tipos de inteiro para double

+++ aula_12_12/RegioesEnum/src/br/com/db1/db1start/Calculadora.java
@@ -1,15 +1,15 @@
 package br.com.db1.db1start;
 
 public class Calculadora {
-	private Integer numero1;
-	private Integer numero2;
+	private Double numero1;
+	private Double numero2;
 	
-	public Calculadora(Integer numero1, Integer numero2) {
+	public Calculadora(double numero1, double numero2) {
 		this.numero1 = numero1;
 		this.numero2 = numero2;
 	}
 	
-	public Integer realizaOperacaoMatematica(Operacao operacao) {
+	public double realizaOperacaoMatematica(Operacao operacao) {
 		return operacao.executor(this.numero1, this.numero2);
 	}
 }
