+++ aula_12_12/RegioesEnum/src/br/com/db1/db1start/Operacao.java
@@ -4,28 +4,28 @@ public enum Operacao {
 	SOMA
 	{
 		@Override
-		public Integer executor(Integer numero1, Integer numero2) {
+		public Double executor(Double numero1, Double numero2) {
 			return numero1 + numero2;
 		}
 	},
  	SUBTRACAO {
 		@Override
-		public Integer executor(Integer numero1, Integer numero2) {
+		public Double executor(Double numero1, Double numero2) {
 			return numero1 - numero2;
 		}
 	},
 	DIVISAO {
 		@Override
-		public Integer executor(Integer numero1, Integer numero2) {
+		public Double executor(Double numero1, Double numero2) {
 			return numero1 / numero2;
 		}
 	},
 	MULTIPLICACAO {
 		@Override
-		public Integer executor(Integer numero1, Integer numero2) {
+		public Double executor(Double numero1, Double numero2) {
 			return numero1 * numero2;
 		}
 	};
 	
-	public abstract Integer executor(Integer numero1, Integer numero2);
+	public abstract Double executor(Double numero1, Double numero2);
 }
